refactor(ED25519/keyloader): Change to load from raw bytes.
The ED25519 key pair was previously loaded from the PEM format.
In the dependency library cryptography, the EDPrivateKey can be used to export in the PEM format,
but there seem to be some issues with the export of the EDPublicKey.
Loading from the original byte string of the ED key pair can solve this problem and is also more convenient.